Include Kernel image and modules in qcow2 image.
The current qcow2 includes the Kernel image and modules from the tar
file that was downloaded from GL-iNet. This means the current virtual
machines have not been using modules (and potentially, the kernel) from
the builds we have been `make`-ing.

This creates a new qcow2 that includes the kernels and modules from a
local `make` named `newbuntu` as instructed at
https://github.com/gl-inet/mv1000-ubuntu-image .

Also you'll note the rm before copying the modules over... the copy -a doesn't
seem to like doing so over existing symlinks.